requirements.txt
django-db-logger
django-constance[database] #http://django-constance.readthedocs.org/en/latest/
django-constance
django-bootstrap3
Pillow